+++ dali/imgcodec/decoders/nvjpeg/nvjpeg.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_IMGCODEC_DECODERS_NVJPEG_H_
-#define DALI_IMGCODEC_DECODERS_NVJPEG_H_
+#ifndef DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG_H_
+#define DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG_H_
 
 #include <nvjpeg.h>
 #include <map>
@@ -114,4 +114,4 @@ class NvJpegDecoderFactory : public ImageDecoderFactory {
 }  // namespace imgcodec
 }  // namespace dali
 
-#endif  // DALI_IMGCODEC_DECODERS_NVJPEG_H_
+#endif  // DALI_IMGCODEC_DECODERS_NVJPEG_NVJPEG_H_
